Se refactorizo el paquete material y sus componentes en producto Issue #13
