+++ application/src/main/java/domainapp/application/manifest/menubars.layout.xml
@@ -34,16 +34,16 @@
             </mb3:section>
         </mb3:menu>
         <mb3:menu>
-            <mb3:named>Materiales</mb3:named>
+            <mb3:named>Productos</mb3:named>
             <mb3:section>
                 <mb3:serviceAction objectType="simple.SimpleMaterialMenu" id="create">
-                    <cpt:named>Crear Nuevo Material</cpt:named>
+                    <cpt:named>Crear Nuevo Producto</cpt:named>
                 </mb3:serviceAction>
                 <mb3:serviceAction objectType="simple.SimpleMaterialMenu" id="findByName">
                     <cpt:named>Buscar por nombre</cpt:named>
                 </mb3:serviceAction>
                 <mb3:serviceAction objectType="simple.SimpleMaterialMenu" id="listAll">
-                    <cpt:named>Listar todos los Materiales</cpt:named>
+                    <cpt:named>Listar todos los Productos</cpt:named>
                 </mb3:serviceAction>
             </mb3:section>
         </mb3:menu>
