+++ module-simple/src/main/java/domainapp/modules/simple/dominio/producto/Producto.layout.xml
@@ -11,7 +11,7 @@
                 <bs3:tab name="General">
                     <bs3:row>
                         <bs3:col span="12">
-                            <c:fieldSet tipo="Datos del material">
+                            <c:fieldSet tipo="Datos del producto">
                                 <c:action id="delete">
                                     <c:describedAs>Deletes this object from the persistent datastore</c:describedAs>
                                 </c:action>
